requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
scikit-learn==1.3.0
apache-airflow==2.7.0
dvc==3.30.1
python-dateutil==2.8.2